+++ slixmpp_omemo/project.py
@@ -4,7 +4,7 @@ project = {
     "name"         : "slixmpp-omemo",
     "description"  : "Slixmpp OMEMO plugin",
     "url"          : "https://github.com/Syndace/slixmpp-omemo",
-    "year"         : "2024",
+    "year"         : "2025",
     "author"       : "Tim Henkes (Syndace)",
     "author_email" : "me@syndace.dev",
     "categories"   : [